remove casting

+++ tpu_inference/layers/vllm/custom_ops/gdn_attention_op.py
@@ -146,10 +146,8 @@ def gdn_attention_core_tpu(
 
         is_active = seq_lens_sliced > 0
         read_state_indices_sliced = jnp.where(is_active,
-                                              read_state_indices_sliced,
-                                              0).astype(jnp.int32)
-        state_indices_sliced = jnp.where(is_active, state_indices_sliced,
-                                         0).astype(jnp.int32)
+                                              read_state_indices_sliced, 0)
+        state_indices_sliced = jnp.where(is_active, state_indices_sliced, 0)
     else:
         # Index mamba state by the per-request slot id from
         # `InputBatch.mamba_state_indices_cpu`, not by `block_tables[:, 0]`
